--- workout-service/src/main/java/com/codecool/enterprise/avatar/model/Exercise.java
@@ -1,5 +1,0 @@
-package com.codecool.enterprise.avatar.model;
-
-public enum Exercise {
-    PULLUP
-}
