+++ workout-service/src/main/java/com/codecool/enterprise/workout/controller/WorkoutController.java
@@ -1,17 +1,11 @@
-package com.codecool.enterprise.avatar.controller;
+package com.codecool.enterprise.workout.controller;
 
-import com.codecool.enterprise.avatar.model.Workout;
-import com.codecool.enterprise.avatar.repository.WorkoutRepository;
-import com.codecool.enterprise.avatar.service.WorkoutService;
-import org.json.JSONArray;
-import org.json.JSONObject;
+import com.codecool.enterprise.workout.model.Workout;
+import com.codecool.enterprise.workout.service.WorkoutService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.SessionAttributes;
 
-import java.io.IOException;
 import java.util.List;
 
 @RestController
