+++ workout-service/src/main/java/com/codecool/enterprise/workout/model/Exercise.java
@@ -1,0 +1,5 @@
+package com.codecool.enterprise.workout.model;
+
+public enum Exercise {
+    PULLUP
+}
