+++ workout-service/src/main/java/com/codecool/enterprise/workout/model/Workout.java
@@ -1,4 +1,4 @@
-package com.codecool.enterprise.avatar.model;
+package com.codecool.enterprise.workout.model;
 
 
 import lombok.AllArgsConstructor;
