+++ workout-service/src/main/java/com/codecool/enterprise/workout/repository/WorkoutRepository.java
@@ -1,9 +1,7 @@
-package com.codecool.enterprise.avatar.repository;
+package com.codecool.enterprise.workout.repository;
 
-import com.codecool.enterprise.avatar.model.Workout;
+import com.codecool.enterprise.workout.model.Workout;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 public interface WorkoutRepository extends JpaRepository<Workout, Long> {
