+++ workout-service/src/main/java/com/codecool/enterprise/workout/service/WorkoutService.java
@@ -1,8 +1,8 @@
-package com.codecool.enterprise.avatar.service;
+package com.codecool.enterprise.workout.service;
 
 
-import com.codecool.enterprise.avatar.model.Workout;
-import com.codecool.enterprise.avatar.repository.WorkoutRepository;
+import com.codecool.enterprise.workout.model.Workout;
+import com.codecool.enterprise.workout.repository.WorkoutRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
